backend/requirements.txt
Django==3.1.5
django-countries==7.0
django-storages[google]==1.11.1
djangorestframework==3.12.2
Pillow==8.1.0
psycopg2-binary==2.8.6
